adaGfxVendor.h: widen Width/Height from uint8_t to uint16_t
Real, previously-undiscovered bug: any display taller/wider than 255px
(e.g. a 240x320 ILI9341) silently truncated instead of failing to compile
— avr-g++ only warns on the narrowing NTTP conversion (-Woverflow), it
doesn't error, so examples/adafruitGfx_MCUFRIEND's own
AdaGfxVendor<320,480,...> had been silently building with
Width=64/Height=224 (320 mod 256, 480 mod 256) the entire time, never
caught. Found while bringing up a real 240x320 ILI9341 on STM32 (arm-gcc
treats the same narrowing as a hard error, surfacing what avr-gcc hid).
kWidth/kHeight and every x/y/w/h coordinate parameter widened to match;
CharW/CharH stay uint8_t (character cell sizes are always small).

+++ include/oneIO/display/adaGfxVendor.h
@@ -23,6 +23,16 @@ namespace oneIO::display {
   // here (matches the already-established "Color<Cor>/Font<Fnt> cascading
   // table" scope, which itself is a 2-state — selected/not — mechanism).
   //
+  // Width/Height (and the x/y/w/h coordinate params below) are uint16_t,
+  // not uint8_t: found 2026-07-25 that any real display taller/wider than
+  // 255px (e.g. a 240x320 ILI9341) silently truncated (320 -> 64 mod 256)
+  // instead of failing to compile — avr-g++ only warns on this
+  // (-Woverflow), it doesn't error, so examples/adafruitGfx_MCUFRIEND's own
+  // AdaGfxVendor<320,480,...> had been silently building with
+  // Width=64/Height=224 the entire time, never caught. uint16_t covers
+  // every real display this wrapper is meant for; CharW/CharH stay uint8_t
+  // (character cell sizes are always small).
+  //
   // flush(): a no-op by default. Most concrete Adafruit_GFX drivers (real
   // SPI/parallel TFTs — ST7735, ILI9341, etc.) push each draw call
   // immediately, no local RAM framebuffer, nothing to flush. This does NOT
@@ -43,11 +53,11 @@ namespace oneIO::display {
   // print simply pick FgColor or BgColor directly based on _inv, no extra
   // setDrawColor-style call needed first.
   /// @brief direct thin wrapper over a real Adafruit_GFX-derived vendor object — no reimplementation, no AM4 dependency
-  template<uint8_t Width, uint8_t Height, uint8_t CharW, uint8_t CharH,
+  template<uint16_t Width, uint16_t Height, uint8_t CharW, uint8_t CharH,
            uint16_t FgColor = 0xFFFF, uint16_t BgColor = 0x0000>
   struct AdaGfxVendor {
-    static constexpr uint8_t kWidth  = Width;
-    static constexpr uint8_t kHeight = Height;
+    static constexpr uint16_t kWidth  = Width;
+    static constexpr uint16_t kHeight = Height;
 
     inline static Adafruit_GFX* driver = nullptr;
     inline static bool          _inv   = false;
@@ -63,7 +73,7 @@ namespace oneIO::display {
     }
     static void print(const char* s) { while (*s) print(*s++); }
 
-    static void setCursor(uint8_t x, uint8_t y) { driver->setCursor(x, y); }
+    static void setCursor(uint16_t x, uint16_t y) { driver->setCursor(x, y); }
     static void setBigFont(bool) {}  // no big/small font toggle — disclosed simplification
     static void clear() { driver->fillScreen(BgColor); }
     static void flush() {}  // see file header comment
@@ -71,10 +81,10 @@ namespace oneIO::display {
 
     // byte (Ssd1306-specific XOR-fill-value) is ignored — the real fill
     // color is always FgColor/BgColor via _inv instead, see header comment.
-    static void fillRect(uint8_t x, uint8_t y, uint8_t w, uint8_t h, uint8_t = 0) {
+    static void fillRect(uint16_t x, uint16_t y, uint16_t w, uint16_t h, uint8_t = 0) {
       driver->fillRect(x, y, w, h, _inv ? FgColor : BgColor);
     }
-    static void drawRoundRect(uint8_t x, uint8_t y, uint8_t w, uint8_t r) {
+    static void drawRoundRect(uint16_t x, uint16_t y, uint16_t w, uint16_t r) {
       driver->fillRoundRect(x, y, w, CharH, r, _inv ? FgColor : BgColor);
     }
 
@@ -91,11 +101,11 @@ namespace oneIO::display {
   // polymorphic base, so display() resolves as an ordinary non-virtual
   // call — same "flush() reaches the real buffered vendor call" shape as
   // U8g2Vendor's own flush()->sendBuffer().
-  template<typename VendorT, uint8_t Width, uint8_t Height, uint8_t CharW, uint8_t CharH,
+  template<typename VendorT, uint16_t Width, uint16_t Height, uint8_t CharW, uint8_t CharH,
            uint16_t FgColor = 0xFFFF, uint16_t BgColor = 0x0000>
   struct AdaGfxBufferedVendor {
-    static constexpr uint8_t kWidth  = Width;
-    static constexpr uint8_t kHeight = Height;
+    static constexpr uint16_t kWidth  = Width;
+    static constexpr uint16_t kHeight = Height;
 
     inline static VendorT* driver = nullptr;
     inline static bool     _inv   = false;
@@ -108,16 +118,16 @@ namespace oneIO::display {
     }
     static void print(const char* s) { while (*s) print(*s++); }
 
-    static void setCursor(uint8_t x, uint8_t y) { driver->setCursor(x, y); }
+    static void setCursor(uint16_t x, uint16_t y) { driver->setCursor(x, y); }
     static void setBigFont(bool) {}
     static void clear() { driver->fillScreen(BgColor); }
     static void flush() { driver->display(); }
     static void setInverted(bool v) { _inv = v; }
 
-    static void fillRect(uint8_t x, uint8_t y, uint8_t w, uint8_t h, uint8_t = 0) {
+    static void fillRect(uint16_t x, uint16_t y, uint16_t w, uint16_t h, uint8_t = 0) {
       driver->fillRect(x, y, w, h, _inv ? FgColor : BgColor);
     }
-    static void drawRoundRect(uint8_t x, uint8_t y, uint8_t w, uint8_t r) {
+    static void drawRoundRect(uint16_t x, uint16_t y, uint16_t w, uint16_t r) {
       driver->fillRoundRect(x, y, w, CharH, r, _inv ? FgColor : BgColor);
     }
